requirements-dev.txt
backports.zoneinfo==0.2.1;python_version<="3.9"
bcrypt
black
build
fastapi
graphene
graphene-mongo
graphql-core
python-dotenv
requests
mongoengine
pyOpenSSL
pytest
setuptools
uvicorn
